src/whisper_inference/requirements.txt
openai_whisper==20230314
pandas==2.0.3
whisper_timestamped==1.12.20
Cython
dtw-python
